moved function to helpers file
